Core Functionalities UI modified
Main Menu presents two options
-> View Attendance
-> Take Attendance

View Attendance:
-> presents a list of courses.

*Not yet implemented*
-> Upon selection of a course will present user with a list of
Attendance Roll sorted by most recent date.

-> Upon selection of a date will present a list of students present for
that date's roll

Take  Attendance:
-> Present users with a list of courses available for that user.
-> Upon selecting a course will launch Scan activity and passing the
course code

Scan Activity:
-> Display Course Code
-> Upon clicking can button launches the barcode scanner
-> Successful scan will display Student ID and presents a confirmation
image
*Not yet implemented*
-> A new roll will be generated using the current date,
-> ID card scanned on that date will be added to that roll
