+++ enroll/app/src/main/res/layout/activity_main.xml
@@ -9,19 +9,31 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     tools:context="com.project.uwi.enroll.MainActivity">
 
-    <Button android:id="@+id/scanid_button"
+    <Button android:id="@+id/take_Att"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
         android:onClick="onClickScan"
-        android:text="Scan ID" />
+        android:text="Take Attendance"
+        android:layout_marginBottom="107dp"
+        android:layout_alignParentBottom="true"
+        android:layout_alignStart="@+id/view_Att" />
 
-    <Button android:id="@+id/view_button"
+    <Button android:id="@+id/view_Att"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:layout_below="@id/scanid_button"
         android:onClick="onClickView"
-        android:text="View" />
+        android:text="View Attendance"
+        android:layout_marginBottom="24dp"
+        android:layout_above="@+id/take_Att"
+        android:layout_centerHorizontal="true" />
+
+    <ImageView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:id="@+id/RR_Logo"
+        android:src="@drawable/logo"
+        android:layout_alignParentTop="true"
+        android:layout_centerHorizontal="true"
+        android:layout_marginTop="31dp" />
 
 </RelativeLayout>
