+++ enroll/app/src/main/res/layout/activity_student_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.project.uwi.enroll.Display">
+    tools:context="com.project.uwi.enroll.StudentList">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
@@ -21,7 +21,7 @@
 
     </android.support.design.widget.AppBarLayout>
 
-    <include layout="@layout/content_display" />
+    <include layout="@layout/content_student_list" />
 
     <android.support.design.widget.FloatingActionButton
         android:id="@+id/fab"
