+++ enroll/app/src/main/res/layout/content_course_list.xml
@@ -9,24 +9,13 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.project.uwi.enroll.Course_Details"
-    tools:showIn="@layout/activity_course__details">
-
-
+    tools:context="com.project.uwi.enroll.CourseList"
+    tools:showIn="@layout/activity_course_list">
 
     <ListView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:id="@+id/listStudents"
-        android:layout_below="@+id/textView2" />
-
-    <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:textAppearance="?android:attr/textAppearanceLarge"
-        android:text="STUDENTS"
-        android:id="@+id/textView2"
+        android:id="@+id/course_list"
         android:layout_alignParentTop="true"
         android:layout_centerHorizontal="true" />
-
 </RelativeLayout>
