+++ enroll/app/src/main/res/layout/content_student_list.xml
@@ -9,22 +9,13 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.project.uwi.enroll.Display"
-    tools:showIn="@layout/activity_display">
+    tools:context="com.project.uwi.enroll.StudentList"
+    tools:showIn="@layout/activity_student_list">
 
     <ListView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:id="@+id/listView"
-        android:layout_below="@+id/textView" />
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textAppearance="?android:attr/textAppearanceLarge"
-        android:text="COURSES"
-        android:id="@+id/textView"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true" />
+        android:id="@+id/listStudents"
+        android:layout_below="@+id/textView2" />
 
 </RelativeLayout>
