+++ enroll/app/src/main/res/values/strings.xml
@@ -1,7 +1,8 @@
 <resources>
-    <string name="app_name">enroll</string>
+    <string name="app_name">UWI Rick Roll</string>
     <string name="title_activity_scan">Scan</string>
     <string name="title_activity_view">View</string>
-    <string name="title_activity_display">Display</string>
-    <string name="title_activity_course__details">Course Details</string>
+    <string name="title_activity_course_list">Select Course</string>
+    <string name="title_activity_student_list">Students</string>
+    <string name="title_activity_display">Select Course</string>
 </resources>
